requirements.txt
contourpy==1.3.1
cycler==0.12.1
Deprecated==1.2.18
fonttools==4.56.0
iniconfig==2.0.0
Jinja2==3.1.5
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
numba==0.61.0
numcodecs==0.15.1
numpy==2.1.3
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
py-modelrunner==0.19.0
py-pde==0.43.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
scipy==1.15.2
six==1.17.0
sympy==1.13.3
tqdm==4.67.1
wrapt==1.17.2
